+++ docs/source/conf.py
@@ -356,6 +356,7 @@ sphinx_gallery_conf = {
             "../../examples/sampling",
             "../../examples/unfolded",
             "../../examples/blind-inverse-problems",
+            "../../examples/co_design",
             "../../examples/self-supervised-learning",
             "../../examples/transforms-equivariance",
             "../../examples/adversarial-learning",